Trying to solve inflections problems with form helpers.
Added stringify_values.rb into spec/support/

Tried setting an inflection to stop rails from converting TodoList to TodosList as the pluralization, instead it should pluralize to TodoLists.

Trying to be able to pass the instance variable into the semantic_form_for block, which is pointing to a non-existent route due to the inflections (I think)

Passing controller specs.